Correct the behaviors of WriteBuffer
WriteBuffer attempt to dereference a pointer without any check. Adds a
safeguard to prevent dereferencing a null pointer to fix it and
nullPointer warning detected by cppcheck.

Besides, WriteBuffer always does hashing and returns true. According to
the reference of FNV-1a for Hash algorithm, the original code uses a
while loop to iterate over the buffer, checking that the current
position is less than the end. WriteBuffer should not do hashing when
the length is 0.